fix(maestro): isolate native feature gates
