+++ .github/workflows/check.yml
@@ -376,15 +376,7 @@ jobs:
           cache-key-suffix: ${{ runner.os }}-${{ runner.arch }}
 
       - name: Clippy
-        run: cargo clippy --workspace -- -D warnings -D clippy::wildcard_imports
-      - name: Check Maestro non-native feature contract
-        env:
-          RUSTFLAGS: "-D warnings"
-        run: |
-          cargo check -p vize_maestro --no-default-features
-          cargo test -p vize_maestro --no-default-features --test non_native_structural
-          cargo check -p vize_maestro --no-default-features --features glyph
-          cargo test -p vize_maestro --no-default-features --features glyph --test non_native_structural
+        run: cargo clippy --workspace -- -D warnings -D clippy::wildcard_imports && bash tools/github/check-maestro-feature-contract.sh
       - name: Install Pkl CLI
         run: npm install --global @pkl-community/pkl@0.27.2
       - name: Test
