+++ crates/vize_maestro/src/ide/corsa_support.rs
@@ -1,33 +1,26 @@
 //! Shared Corsa helpers for mapping virtual document responses back to Vue SFCs.
+#![cfg(feature = "native")]
 #![allow(clippy::disallowed_types, clippy::disallowed_methods)]
 
 use tower_lsp::lsp_types::{Location, PrepareRenameResponse, Range, Url};
-#[cfg(feature = "native")]
 mod canonical;
-#[cfg(all(test, feature = "native"))]
+#[cfg(test)]
 mod canonical_tests;
-#[cfg(feature = "native")]
 mod html_attribute;
-#[cfg(all(test, feature = "native"))]
+#[cfg(test)]
 mod html_attribute_tests;
-#[cfg(feature = "native")]
 mod html_tag;
-#[cfg(feature = "native")]
 mod svg_attribute;
-#[cfg(feature = "native")]
 mod virtual_mirror;
 mod workspace_edit;
 
-#[cfg(feature = "native")]
 pub(crate) use canonical::{
     canonical_source_offset_to_position, map_canonical_corsa_locations, map_canonical_lsp_range,
     open_canonical_virtual_document,
 };
-#[cfg(feature = "native")]
 pub(crate) use html_attribute::{
     html_attribute_request_path, html_attribute_virtual_document, native_dom_attribute_info,
 };
-#[cfg(feature = "native")]
 pub(crate) use html_tag::{html_tag_request_path, html_tag_virtual_document, native_dom_tag_info};
 pub(crate) use workspace_edit::map_corsa_workspace_edit;
